djosify/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djoser==2.3.1
djangorestframework_simplejwt==5.4.0
python-dotenv==1.0.1
drf-yasg==1.21.8
psycopg2-binary==2.9.10
